refactor: extract StopConditions and TopologicalOrderIter to traversal.rs
DAG traversal is a distinct concern from actor execution. Moving these
types out of executor.rs into their own module clarifies the separation.
